back-end/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
aiohttp==3.12.13
python-dotenv==1.0.1
beautifulsoup4==4.12.3
